+++ fj-doc-playground-quarkus/src/main/react/src/playground/DocXmlEditor.jsx
@@ -187,6 +187,19 @@ class DocXmlEditor extends Component {
 			} else if ( this.state.docFormat === 'MD' ) {
 				let decodedStringAtoB = this.myAtob(this.state.docOutput);
 				outputData = <MarkdownEditor  value={decodedStringAtoB} />
+			} else if ( this.state.docFormat === 'XML' ) {
+				let decodedStringAtoB = this.myAtob(this.state.docOutput);
+				outputData = <AceEditor
+							mode='xml'
+							theme="xcode"
+							name="DOC_XML_OUTPUT"
+							editorProps={{ $blockScrolling: true }}
+							enableBasicAutocompletion={true}
+							enableLiveAutocompletion={true}
+							enableSnippets={true}
+							value={decodedStringAtoB}
+							width='100%'
+						/>
 			}
 			outputData = <Fragment>{outputData}<p>Generation time : {this.state.generationTime}</p></Fragment>
 		} else if ( this.state.outputMessage != null ) {
@@ -244,6 +257,7 @@ class DocXmlEditor extends Component {
 					    <MenuItem value='PDFA'>PDF/A</MenuItem>
 					    <MenuItem value='XLSX'>XLSX</MenuItem>
 					    <MenuItem value='MD'>Markdown (MD)</MenuItem>
+					     <MenuItem value='XML'>Venus XML Doc</MenuItem>
 					  </Select>
 					</FormControl>	
 			  </Grid>